+++ src/main/java/com/project/homes/entity/House.java
@@ -3,6 +3,7 @@ package com.project.homes.entity;
 import lombok.Data;
 import org.springframework.data.mongodb.core.mapping.Document;
 
+import java.math.BigDecimal;
 import java.util.Objects;
 
 @Data
@@ -11,7 +12,7 @@ public class House extends Property {
     private Boolean hasBasement;
     private int numOfFloors;
 
-    public House(PropertyType type, Double area, Double pricing, Address address, int numOfRooms, int numOfBaths,
+    public House(PropertyType type, Double area, BigDecimal pricing, Address address, int numOfRooms, int numOfBaths,
                  String pictureUrl, String ownerId, Boolean hasBasement, int numOfFloors) {
         super(type, area, pricing, address, numOfRooms, numOfBaths, pictureUrl, ownerId);
         this.hasBasement = hasBasement;
